Shader Abstraction and Uniforms

+++ Axis/src/Axis/Nuklear/nuklear_glfw_opengl_impl.h
@@ -341,7 +341,7 @@ nk_gflw3_resize_callback(GLFWwindow* win, int width, int height)
 
     int w, h;
     glfwGetFramebufferSize(win, &w, &h);
-    nk_dockspace_adjust(&glfw->ctx, w, h);
+    nk_dockspace_adjust(&glfw->ctx, (float)w, (float)h);
 }
 
 NK_API void
